tidy up

+++ Klasa-2/Lekcje/11_Kontest/neuronow/dij.cpp
@@ -6,8 +6,8 @@ using namespace std;
 
 const int MAXN = 1e5 + 7;
 vector<int> power_plants(MAXN);
-vector<pair<int, int>> graph[MAXN];
-vector<int> dist(MAXN, -1);
+vector<pair<__int128, __int128>> graph[MAXN];
+vector<__int128> dist(MAXN, -1);
 
 void dijkstra(int s) {
     priority_queue<pair<int, int>, vector<pair<int, int>>, greater<pair<int, int>>> pq;
@@ -43,7 +43,7 @@ int main() {
 
     dijkstra(e + d);
 
-    int minx = dist[1];
+    __int128 minx = dist[1];
     int mini = 1;
     for (int i = 1; i <= e; i++) {
         if (dist[i] < minx) {
